Function libmake::args::process_arguments

pub fn process_arguments(matches: &ArgMatches) -> Result<(), Box<dyn Error>>

Processes the command line arguments provided to the program.

This function takes in command-line arguments parsed by clap and performs actions based on the arguments provided.

Arguments

matches - An instance of clap::ArgMatches containing the parsed command line arguments.

Errors

This function will return an error if there is an issue with processing the command line arguments or generating files.