Function libmake::cli::build

pub fn build() -> Result<ArgMatches, Error>

Constructs the command-line interface for the application using Clap, including all necessary arguments.

Examples

use libmake::cli;
let matches = cli::build().expect("CLI parsing failed");

Errors

This function will return an error if the command-line argument parsing fails.