Function libmake::run

pub fn run() -> Result<(), Box<dyn Error>>

Initializes the logger with a file logger and a terminal logger and processes command-line arguments to generate the new library.

Errors

This function will return an error in the following situations:

If there is a problem generating ASCII art for the tool’s CLI.

If an error occurs while building the command-line interface or processing arguments.

Any other errors that may arise from operations performed within the function.