Module libmake::generator

The generator module contains functions for generating the new library.

Structs

FileGenerationParams

Structure for holding the parameters for generating the project files.

Functions

copy_and_replace_template

Copies a template file to the output directory and replaces the placeholders with the given parameters (if any).

create_directory

Creates a directory if it does not exist. If the directory already exists, the function will return Ok(()).

create_template_folder

Creates the template directory and downloads the template files.

generate_files

Generates files for a new Rust project based on given arguments.

generate_from_args

Generates files for a new Rust project based on command line arguments. The arguments must be in the form --name=value. The following arguments are supported:

generate_from_config

Generates files for a new Rust project based on a configuration file. Supported configuration file formats: JSON, YAML, CSV.

generate_from_csv

Generates files for a new Rust project based on a CSV file.

generate_from_json

Generates files for a new Rust project based on a JSON file.

generate_from_toml

Generates files for a new Rust project based on a TOML file.

generate_from_yaml

Generates files for a new Rust project based on a YAML file.