libmake::generator

Function copy_and_replace_template

pub fn copy_and_replace_template(
    template_file: &str,
    dest_file: &str,
    project_directory: &PathBuf,
    params: &FileGenerationParams,
) -> Result<()>

Copies a template file to the output directory and replaces the placeholders with the given parameters (if any).

Arguments

template_file - The name of the template file.

dest_file - The name of the destination file.

project_directory - The path to the project directory.

params - The parameters to be used for replacing the placeholders.

Errors

Returns an io::Error in the following cases:

If the template file cannot be found, read, or copied.

If the destination file cannot be created or written to.

If there is an error in replacing placeholders in the destination file.