libmake::interface

Function replace_placeholders

pub fn replace_placeholders(
    template_file: &PathBuf,
    output_file: &PathBuf,
    params: &FileGenerationParams,
) -> Result<()>

Replaces placeholders in a template file with values from the provided parameters and writes the result to an output file.

Arguments

template_file - Path to the template file.

output_file - Path to the output file where the result should be written.

params - Parameters containing values to replace placeholders in the template.

Errors

Returns an std::io::Result error if:

The template file cannot be read.

The output file cannot be created or written to.

There are issues parsing the template or replacing placeholders.