libmake

Macro macro_ascii

macro_rules! macro_ascii {
    ($text:expr) => { ... };
}

A macro for generating ASCII art from text.

This macro takes a string literal as input and generates ASCII art using the generate_ascii_art function. If the conversion is successful, the macro returns the ASCII art as a string. If an error occurs during the conversion, the macro panics with an error message.

Examples

use libmake::macro_ascii;

let art = macro_ascii!("Hello, world!");
println!("{}", art);