Macro libmake::macro_replace_placeholder

macro_rules! macro_replace_placeholder {
    ($line:expr, $params:expr, $($field:ident),+) => { ... };
}

Replaces placeholders in a given line with corresponding values from the provided parameters.

Arguments

line - The line containing placeholders to be replaced.

params - The parameters containing values to replace the placeholders.

$($field:ident),+ - Identifiers representing the fields in params to be replaced.

Returns

The line with placeholders replaced by their corresponding values.