libmake::macros

Module directory_macros

The directory_macros module contains macros related to directory operations. This module provides macros for directory operations, including checking directory existence, creating multiple directories at once, and cleaning up directories.

macro_check_directory Macro

Checks if a directory exists and creates it if necessary.

Usage

use libmake::macro_check_directory;
use std::path::Path;

let path = Path::new("logs");
macro_check_directory!(path, "logs");

Arguments

_dir - The path to check, as a std::path::Path.

_name - A string literal representing the directory name. This is used in error messages.

Behaviour

The macro_check_directory macro checks if the directory specified by _dir exists. If it exists and is not a directory, a panic with an error message is triggered. If the directory doesn’t exist, the macro attempts to create it using std::fs::create_dir_all(_dir).

macro_cleanup_directories Macro

Cleans up multiple directories by invoking the cleanup_directory function.

Usage

use std::path::Path;
use libmake::macro_check_directory;

let path = Path::new("logs");
macro_check_directory!(path, "logs");

Arguments

$( $_dir:expr ),* - A comma-separated list of directory paths to clean up.

Behaviour

The macro_cleanup_directories macro takes multiple directory paths as arguments and invokes the cleanup_directory function for each path.

macro_create_directories Macro

Creates multiple directories at once.

Usage

use libmake::{macro_create_directories, macro_cleanup_directories};
use std::path::Path;

macro_create_directories!("logs", "logs1", "logs2");
macro_cleanup_directories!(Path::new("./logs"), Path::new("./logs1"), Path::new("./logs2"));

Arguments

... - Variable number of directory paths, each specified as an expression (expr).

Behaviour

The macro_create_directories macro creates multiple directories at once.

Example

use libmake::{macro_create_directories, macro_cleanup_directories};
use std::path::Path;

fn main() -> Result<(), Box<dyn std::error::Error>> {
    let test = Path::new("logs");
    let test2  = Path::new("logs1");
    macro_create_directories!(test, test2)?;
    macro_cleanup_directories!(test, test2);
    Ok(())
}

Note

These macros assume familiarity with Rust macros and their usage. Users are encouraged to review Rust macro documentation for a better understanding of how to work with macros effectively.