libmake::utilities::directory

Function move_output_directory

pub fn move_output_directory(site_name: &str, out_dir: &Path) -> Result<()>

Moves the output directory to the public directory.

This function takes a reference to a Path object for the output directory and a string for the site name, and moves the output directory to the public directory.

Arguments

site_name - A string for the site name.

out_dir - A reference to a Path object for the output directory.

Returns

Result<(), std::io::Error> - A result indicating success or failure.

Ok(()) if the output directory was moved successfully.

Err(std::io::Error) if the output directory could not be moved.