Module libmake::utilities::directory

The directory module contains functions related to directory operations.

Functions

cleanup_directory

Cleans up the directory at the given path.

create_directory

Creates a new directory at the given path.

directory

Ensures a directory exists, creating it if necessary.

move_output_directory

Moves the output directory to the public directory.

truncate

Truncates a path to only have a set number of path components.